requirements.txt
appdirs==1.4.4
backcall==0.2.0
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cssselect==1.1.0
decorator==4.4.2
fake-useragent==0.1.11
gitdb==4.0.5
GitPython==3.1.11
idna==2.10
ipykernel==5.3.4
ipython==7.19.0
ipython-genutils==0.2.0
jedi==0.17.2
jupyter-client==6.1.7
jupyter-core==4.7.0
lxml==4.6.2
parse==1.18.0
parso==0.7.1
pickleshare==0.7.5
prompt-toolkit==3.0.8
pyee==7.0.4
Pygments==2.7.2
pyppeteer==0.2.2
pyquery==1.4.3
python-dateutil==2.8.1
pywin32==300
pyzmq==20.0.0
requests==2.25.0
requests-html==0.10.0
six==1.15.0
smmap==3.0.4
soupsieve==2.0.1
tornado==6.1
tqdm==4.54.0
traitlets==5.0.5
urllib3==1.26.2
w3lib==1.22.0
wcwidth==0.2.5
websockets==8.1
